requirements.txt
aiogram==3.4.1
netmiko==4.3.0
PyYAML==6.0.1
